% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualize_model.R
\name{actualize_model}
\alias{actualize_model}
\title{Update MAP estimations across occasions for each IDs}
\usage{
actualize_model(
  actualization_map,
  evaluation_type = c("Progressive", "Most_Recent_Progressive", "Cronologic_Ref",
    "Most_Recent_Ref")
)
}
\arguments{
\item{actualization_map}{A list containing the MAP estimations obtained in `run_MAP_estimations`
Must include an element named `map_estimations` which stores the posterior estimations
for each occasion.}

\item{evaluation_type}{A character vector specifying the evaluation type to use for updating.
Options include:
- `"Progressive"`: Use posterior results progressively across all previous occasions.
- `"Most_Recent_Progressive"`: Use only the most recent posterior for updating.
- `"Cronologic_Ref"`: Use a chronological reference for posterior updates.
- `"Most_Recent_Ref"`: Use the most recent chronological reference for updates.
Defaults to `"Progressive"`.}
}
\value{
A list with two elements:
  - `ind_model`: A list containing posterior estimations (`a.posteriori`) for each occasion.
  - `eval_type`: The evaluation type used for the process.
}
\description{
The `actualize_model` function updates MAP (Maximum A Posteriori) estimations
across occasions based on a selected evaluation type. It processes posterior estimations
using the specified strategy and returns the updated model for each ID on every single OCC
}
\details{
This function evaluates posterior estimations iteratively based on the specified
`evaluation_type`. It ensures compatibility with the evaluation type used during
the creation of the input `actualization_map`.

The function dynamically names the posterior estimation results, following the pattern
`a.posteriori_occX_Y`, where `X` and `Y` represent the occasions used in the estimation.
}
\examples{
# Example input data
actualization_map <- list(
  eval_type = "Progressive",
  map_estimations = list(
    "map.estimation.occ_0_1" = NULL,
    "map.estimation.occ_0_1_2" = list() # Replace with real MAP estimation object
  )
)

# Run the function
result <- actualize_model(actualization_map, evaluation_type = "Progressive")
print(result)

}
