% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tacHAN2011.R
\docType{data}
\name{data_tacHAN2011}
\alias{data_tacHAN2011}
\title{Tacrolimus data from Han et al. (2011)}
\format{
A tibble: 739 rows and 30 variables
}
\source{
Han, Nayoung, et al. "Prediction of the tacrolimus population pharmacokinetic parameters according to CYP3A5 genotype and clinical factors using NONMEM in adult kidney transplant recipients." European journal of clinical pharmacology 69.1 (2013): 53-63.
}
\usage{
data(data_tacHAN2011)
}
\description{
Tacrolimus data from Han et al. (2011)
}
\keyword{datasets}
