% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exeval_ppk.R
\name{exeval_ppk}
\alias{exeval_ppk}
\title{External evaluation for Population Pharmacokinetic (popPK) Models}
\usage{
exeval_ppk(
  model_name,
  drug_name,
  model_code,
  tool = "mapbayr",
  check_compile = TRUE,
  data,
  num_occ = NULL,
  num_ids = NULL,
  sampling = TRUE,
  occ_ref = NULL,
  evaluation_type = c("Progressive", "Most_Recent_Progressive", "Cronologic_Ref",
    "Most_Recent_Ref"),
  names_occ = "OCC",
  names_id = "ID",
  names_time = "TIME",
  names_evid = "EVID",
  method = c("L-BFGS-B", "newuoa"),
  assessment = c("a_priori", "Bayesian_forecasting", "Complete"),
  verbose = FALSE
)
}
\arguments{
\item{model_name}{Character string. Name of the model to use in the analysis.}

\item{drug_name}{Character string. Used only for reporting purposes.}

\item{model_code}{Character string. Code of the pharmacokinetic model in mrgsolve format.}

\item{tool}{Character string. Specifies the tool to use for estimation. Currently "mapbayr" is the only option.}

\item{check_compile}{Logical. If `TRUE`, checks if the model compiles correctly in `mapbayr`.}

\item{data}{Data frame. Contains the input data for the estimations, including columns like ID, TIME, and OCC.}

\item{num_occ}{Integer. Number of occasions (OCC) to include in the analysis. If `NULL`, all unique occasions in `data` are used.}

\item{num_ids}{Integer. Number of unique IDs to include in the analysis. If `NULL`, all IDs are included.}

\item{sampling}{Logical. If `TRUE`, randomly samples the specified number of IDs from the data.}

\item{occ_ref}{Integer. Reference occasion for evaluation types that require a reference. Must be consistent with `evaluation_type`.}

\item{evaluation_type}{Character string. Specifies the evaluation type. Options are:
\itemize{
  \item "Progressive": Uses all data up to each occasion.
  \item "Most_Recent_Progressive": Uses only the most recent occasion.
  \item "Cronologic_Ref": Uses all data up to a reference occasion.
  \item "Most_Recent_Ref": Uses the most recent occasion relative to a reference.
}}

\item{names_occ}{Character string. Name of the OCC column in the data. Defaults to "OCC".}

\item{names_id}{Character string. Name of the ID column in the data. Defaults to "ID".}

\item{names_time}{Character string. Name of the TIME column in the data. Defaults to "TIME".}

\item{names_evid}{Character string. Name of the EVID column in the data. Defaults to "EVID".}

\item{method}{Character vector. Specifies optimization methods for `mapbayr`. Options are "L-BFGS-B" or "newuoa".}

\item{assessment}{Character string. Specifies the type of prediction to perform. Options are:
\itemize{
  \item "a_priori": Simulates concentrations using the population model without individual data.
  \item "Bayesian_Forecasting": Simulates concentrations using individual parameter estimates (posterior mode).
  \item "Complete": Performs both a priori and Bayesian forecasting simulations.
}}

\item{verbose}{Logical. If TRUE, messages are printed during execution.
If FALSE (default), errors are stored as warnings accessible with `warnings()`}
}
\value{
A list containing:
\describe{
  \item{metrics}{Evaluation metrics}
  \item{estimations}{MAP estimation results for each subset of the data.}
  \item{simulation}{A list of simulation results for each occasion and individual.}
  \item{updates}{A list containing posterior estimations (`a.posteriori`) for each occasion.}
}
}
\description{
Performs external evaluation of popPK models by conducting MAP estimations and individual simulations
for each occasion using different evaluation strategies (see evaluation_type)
}
\examples{
\dontrun{
data("data_tacHAN2011", package = "preDose")
data("model_tacHAN2011", package = "preDose")

dd <- data_tacHAN2011 |> subset(ID < 6)

res <- exeval_ppk(model_name = "tacrolimus_HAN2011",
                 model_code = model_tacHAN2011,
                 data = dd,
                 evaluation_type= "Progressive",
                 assessment='Bayesian_forecasting' )

res # Print the results
}
}
