% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_data_mapbayr.R
\docType{data}
\name{external_data_mapbayr}
\alias{external_data_mapbayr}
\title{Tacrolimus data from Han et al. (2011) (DUPLICATED)}
\format{
A tibble: 739 rows and 30 variables
}
\source{
Han, Nayoung, et al. "Prediction of the tacrolimus population pharmacokinetic parameters according to CYP3A5 genotype and clinical factors using NONMEM in adult kidney transplant recipients." European journal of clinical pharmacology 69.1 (2013): 53-63.
}
\usage{
data(external_data_mapbayr)
}
\description{
Tacrolimus data from Han et al. (2011) (DUPLICATED)
}
\keyword{datasets}
