% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_occ.R
\name{metrics_occ}
\alias{metrics_occ}
\title{Compute pharmacokinetic evaluation metrics by occasion}
\usage{
metrics_occ(x, ...)
}
\arguments{
\item{x}{An object containing individual simulations.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
An object of class \code{EvalMetricsPPK}.
}
\description{
Generic function to compute prediction error metrics by occasion (OCC).
The method dispatched depends on the class of \code{x}.
}
