% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_plot.EvalMetricsPPK.R
\name{metrics_plot.EvalMetricsPPK}
\alias{metrics_plot.EvalMetricsPPK}
\title{metrics_Plot}
\usage{
\method{metrics_plot}{EvalMetricsPPK}(
  x,
  type = c("bias_barplot", "bias_pointrange", "MAIPE_barplot", "bias_boxplot",
    "bias_violin", "bias_dotplot", "bias_density", "IF20_plot", "IF30_plot", "IF_plot",
    "error_plot"),
  ...
)
}
\arguments{
\item{type}{A character string specifying the type of plot to generate. Options are:
\itemize{
  \item \code{"bias_barplot"}: Bar plot of relative bias (\code{rBIAS}) with error bars.
  \item \code{"bias_pointrange"}: pointrange for rBIAS.
  \item \code{"MAIPE_barplot"}: Bar plot of MAIPE values.
  \item \code{"bias_boxplot"}: Box plot of IPE values.
  \item \code{"bias_dotplot"}: Dotplot of rBIAS values. Variability on individual bias
  \item \code{"bias_density"}: Density Plot for rBias throughout occasions .
  \item \code{"bias_violin"}: Violin plot of IPE values.
  \item \code{"IF20_plot"}: Bar plot of IF20 values with reference line at 35%.
  \item \code{"IF30_plot"}: Bar plot of IF30 values with reference line at 50%.
  \item \code{"IF_plot"}: Combine both IF20 and IF30 plots.
  \item \code{"error_plot"}: Stacked bar plot showing the proportion of prediction errors within predefined IPE bands.
}}

\item{mm}{A list containing data frames with the required  metrics. Typically, `mm[[1]]` and `mm[[2]]` contain relevant data. Values comes from results from [metrics_occ()] function}
}
\value{
A ggplot object corresponding to the selected plot type.
}
\description{
Generate Different Type of Metrics Plots
}
\details{
This function creates various types of metrics plots for evaluating popPK Models , such as rBias, MAIPE, and IF20/IF30 values.


The function utilizes ggplot2 for visualization and `scale_fill_brewer(palette = "Dark2")` for consistent color schemes.
}
