% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Plot.R
\name{metrics_plot}
\alias{metrics_plot}
\title{Plot pharmacokinetic evaluation metrics}
\usage{
metrics_plot(x, ...)
}
\arguments{
\item{x}{Object containing evaluation metrics}

\item{...}{Additional arguments}
}
\description{
Plot pharmacokinetic evaluation metrics
}
