% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tacHAN2011.R
\docType{data}
\name{model_tacHAN2011}
\alias{model_tacHAN2011}
\title{Tacrolimus model from Han et al. (2011)}
\format{
Character describing the PPK model in mrgsolve format.
}
\source{
Han, Nayoung, et al. "Prediction of the tacrolimus population pharmacokinetic parameters according to CYP3A5 genotype and clinical factors using NONMEM in adult kidney transplant recipients." European journal of clinical pharmacology 69.1 (2013): 53-63.
}
\usage{
data(model_tacHAN2011)
}
\description{
Tacrolimus model from Han et al. (2011)
}
\keyword{datasets}
