% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{plot_combined}
\alias{plot_combined}
\title{Plot Combined Metrics}
\usage{
plot_combined(
  cmetrics,
  type = c("bias_barplot", "MAIPE_barplot", "IF20_plot", "IF30_plot")
)
}
\arguments{
\item{cmetrics}{A dataframe obtained from `combine_metrics()`, containing the combined metrics data.}

\item{type}{A character string specifying the type of plot to generate. Options are:
\itemize{
  \item \code{"bias_barplot"}: Bar plot of relative bias (\code{rBIAS}) with error bars.
  \item \code{"MAIPE_barplot"}: Bar plot of MAIPE values.
  \item \code{"IF20_plot"}: Bar plot of IF20 values with reference line at 35%.
  \item \code{"IF30_plot"}: Bar plot of IF30 values with reference line at 50%.
}}
}
\value{
A `ggplot2` object representing the selected plot.
}
\description{
This function generates different types of metrics plots based on the combined metrics from multiple models.
}
\examples{
#' set.seed(123)  # Para reproducibilidad
generate_fake_metrics <- function(n_occasions = 3) {
data.frame(
OCC = rep(1:n_occasions),  # Simula varias ocasiones
rBIAS = rnorm(n_occasions, mean = 0, sd = 10),
rBIAS_lower = rnorm(n_occasions, mean = -5, sd = 5),
rBIAS_upper = rnorm(n_occasions, mean = 5, sd = 5),
MAIPE = runif(n_occasions, min = 10, max = 50),
IF20 = runif(n_occasions, min = 20, max = 80),
IF30 = runif(n_occasions, min = 30, max = 90)
)
}
# Save Results of metrics
simulation1 <- list(metrics_means = generate_fake_metrics())
simulation2 <- list(metrics_means = generate_fake_metrics())
# List of models
models_list <- list(
list(model_name = "Test_Model1", metrics_list = simulation1),
list(model_name = "Test_Model2", metrics_list = simulation2)
)
combined_results <- combine_metrics(models_list)
combined_results <- combine_metrics(models_list)
plot_combined(combined_results, type = 'bias_barplot')
}
