% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.EvalPPK.R
\name{print.EvalPPK}
\alias{print.EvalPPK}
\title{print method for EvalPPK class}
\usage{
\method{print}{EvalPPK}(x, ...)
}
\arguments{
\item{x}{EvalPPK object}

\item{...}{additional arguments (not used)}
}
\description{
print method for EvalPPK class
}
