% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MAP_estimations.R
\name{run_MAP_estimations}
\alias{run_MAP_estimations}
\title{Run MAP Estimations}
\usage{
run_MAP_estimations(
  model_name,
  model_code,
  tool = "mapbayr",
  check_compile = TRUE,
  data,
  num_occ = NULL,
  num_ids = NULL,
  sampling = TRUE,
  occ_ref = NULL,
  evaluation_type = c("Progressive", "Most_Recent_Progressive", "Cronologic_Ref",
    "Most_Recent_Ref"),
  names_occ = "OCC",
  names_id = "ID",
  names_time = "TIME",
  names_evid = "EVID",
  method = c("L-BFGS-B", "newuoa")
)
}
\arguments{
\item{model_name}{Character string. Name of the model to use in the analysis.}

\item{model_code}{Character string. Code of the pharmacokinetic model in mrgsolve format.}

\item{tool}{Character string. Specifies the tool to use for estimation. Currently using "mapbayr".}

\item{check_compile}{Logical. If `TRUE`, checks if the model compiles correctly in `mapbayr`.}

\item{data}{Data frame. Contains the input data for the estimations, including columns like ID, TIME, and OCC.}

\item{num_occ}{Integer. Number of occasions (OCC) to include in the analysis. If `NULL`, all unique occasions in `data` are used.}

\item{num_ids}{Integer. Number of unique IDs to include in the analysis. If `NULL`, all IDs are included.}

\item{sampling}{Logical. If `TRUE`, randomly samples the specified number of IDs from the data.}

\item{occ_ref}{Integer. Reference occasion for evaluation types that require a reference. Must be consistent with `evaluation_type`.}

\item{evaluation_type}{Character string. Specifies the evaluation type. Options are:
\itemize{
  \item "Progressive": Uses all data up to each occasion.
  \item "Most_Recent_Progressive": Uses only the most recent occasion.
  \item "Cronologic_Ref": Uses all data up to a reference occasion.
  \item "Most_Recent_Ref": Uses the most recent occasion relative to a reference.
}}

\item{names_occ}{Character string. Name of the OCC column in the data. Defaults to "OCC".}

\item{names_id}{Character string. Name of the ID column in the data. Defaults to "ID".}

\item{names_time}{Character string. Name of the TIME column in the data. Defaults to "TIME".}

\item{names_evid}{Character string. Name of the EVID column in the data. Defaults to "EVID".}

\item{method}{Character vector. Specifies optimization methods for `mapbayr`. Options are "L-BFGS-B" or "newuoa".}
}
\value{
A list containing:
\describe{
  \item{data_by_occ}{Filtered datasets by occasion.}
  \item{treatments_by_occ}{List of treatments grouped by occasion.}
  \item{treatments_apriori}{List of "a priori" treatments }
  \item{map_estimations}{MAP estimation results for each subset of the data.}
  \item{eval_type}{The evaluation type used.}
  \item{pop_model}{The population PK or PKPD model}
}
}
\description{
Performs Maximum A Posteriori (MAP) estimations for pharmacokinetic models
using either `mapbayr` or `lixoftConnectors`.
}
\examples{
\dontrun{
# Example using mapbayr
run_MAP_estimations(
  model_name = "example_model",
  model_code = "example_code",
  tool = "mapbayr",
  data = example_data,
  evaluation_type = "Progressive"
)
}
}
