% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best_models.R
\name{select_best_models}
\alias{select_best_models}
\title{seect_best_models}
\usage{
select_best_models(
  data,
  metric,
  top_n = 3,
  occ_eval = NULL,
  rank_criteria = "min"
)
}
\arguments{
\item{data}{A dataframe obtained from `combine_metrics()`, containing the combined metrics data.}

\item{metric}{A character string specifying the metric to use for ranking.}

\item{top_n}{An integer specifying the number of top models to select per OCC.}

\item{occ_eval}{A numeric or character value specifying the `OCC` to filter the models by. Default is `NULL`, which means no filtering by `OCC`.}

\item{rank_criteria}{A character string specifying how to rank the models:
- `'min'`: Select models with the lowest metric values (e.g., MAIPE).
- `'max'`: Select models whose metric values are higher (e.g., IF30).
- `'abs'` : Selects models with the smallest absolute metric values (eg rBIAS)}
}
\value{
A data frame with the top-ranked models for each OCC.
}
\description{
Select Top Ranked Models
}
\details{
#' This function selects the best models from a dataframe of metrics by models based on a specified ranking metric
(e.g., minimum, maximum, or absolute value of a given metric). It can filter the best models by the
`OCC` value if specified, or return the top models for all `OCC` values.
}
\examples{
top_models <- select_top_models(combined_results, metric = 'MAIPE', top_n = 3, occ_eval=2, rank_criteria = 'min')
}
