% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_Plot.R
\name{metrics_Plot}
\alias{metrics_Plot}
\title{metrics_Plot}
\usage{
metrics_Plot(
  mm,
  type = c("bias_barplot", "MAIPE_barplot", "bias_boxplot", "bias_violin",
    "bias_dotplot", "bias_density", "IF20_plot", "IF30_plot")
)
}
\arguments{
\item{mm}{A list containing data frames with the required  metrics. Typically, `mm[[1]]` and `mm[[2]]` contain relevant data. Values comes from results from [metrics_occ()] function}

\item{type}{A character string specifying the type of plot to generate. Options are:
\itemize{
  \item \code{"bias_barplot"}: Bar plot of relative bias (\code{rBIAS}) with error bars.
  \item \code{"MAIPE_barplot"}: Bar plot of MAIPE values.
  \item \code{"bias_boxplot"}: Box plot of IPE values.
  \item \code{"bias_dotplot"}: Dotplot of rBIAS values. Variability on individual bias
  \item \code{"bias_density"}: Density Plot for rBias throughout occasions .
  \item \code{"bias_violin"}: Violin plot of IPE values.
  \item \code{"IF20_plot"}: Bar plot of IF20 values with reference line at 35%.
  \item \code{"IF30_plot"}: Bar plot of IF30 values with reference line at 50%.
}}
}
\value{
A ggplot object corresponding to the selected plot type.
}
\description{
Generate Different Type of Metrics Plots
}
\details{
This function creates various types of metrics plots for evaluating popPK Models , such as rBias, MAIPE, and IF20/IF30 values.


The function utilizes ggplot2 for visualization and `scale_fill_brewer(palette = "Dark2")` for consistent color schemes.
}
\examples{
\dontrun{
data_list <- list(
  data.frame(OCC = rep(1:3, each = 10), IPE = rnorm(30)),
  data.frame(OCC = rep(1:3, each = 10), rBIAS = rnorm(30, 0, 10),
             rBIAS_lower = rnorm(30, -5, 5), rBIAS_upper = rnorm(30, 5, 5),
             MAIPE = rnorm(30, 20, 5), IF20 = runif(30, 20, 40),
             IF30 = runif(30, 30, 60))
)
plot <- metrics_Plot(data_list, type = "bias_barplot")
print(plot)
}
}
