% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{plot_combined}
\alias{plot_combined}
\title{Plot Combined Metrics}
\usage{
plot_combined(
  cmetrics,
  type = c("bias_barplot", "MAIPE_barplot", "IF20_plot", "IF30_plot")
)
}
\arguments{
\item{cmetrics}{A dataframe obtained from `combine_metrics()`, containing the combined metrics data.}

\item{type}{A character string specifying the type of plot to generate. Options are:
\itemize{
  \item \code{"bias_barplot"}: Bar plot of relative bias (\code{rBIAS}) with error bars.
  \item \code{"MAIPE_barplot"}: Bar plot of MAIPE values.
  \item \code{"IF20_plot"}: Bar plot of IF20 values with reference line at 35%.
  \item \code{"IF30_plot"}: Bar plot of IF30 values with reference line at 50%.
}}
}
\value{
A `ggplot2` object representing the selected plot.
}
\description{
This function generates different types of metrics plots based on the combined metrics from multiple models.
}
\examples{
combined_results <- combine_metrics(models_list)
plot_combined(combined_results, type = 'bias_barplot')
}
